@@ -5,15 +5,15 @@ homepage: https://github.com/Vanethos/flutter_connectivity_widget/
 issue_tracker: https://github.com/Vanethos/flutter_connectivity_widget/issues
 
 environment:
-  sdk: ">=2.2.2 <3.0.0"
+  sdk: ">=2.12.0 <3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  rxdart: ^0.26.0
-  connectivity: ^3.0.2
+  rxdart: ^0.27.2
+  connectivity: ^3.0.6
   stream_disposable: ^0.1.0
-  http: ^0.13.0
+  http: ^0.13.4
 
 dev_dependencies:
   flutter_test:
